Symfony Final Project Finished

+++ u6-symfony/symfony-final-project/src/Controller/SongController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/song')]
 class SongController extends AbstractController
@@ -25,6 +26,11 @@ class SongController extends AbstractController
     #[Route('/new', name: 'app_song_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            // Redirige a la página deseada para usuarios que no tienen el rol ROLE_ADMIN
+            return $this->redirectToRoute('app_denied');
+        }
+        
         $song = new Song();
         $form = $this->createForm(SongType::class, $song);
         $form->handleRequest($request);
@@ -63,16 +69,28 @@ class SongController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_song_show', methods: ['GET'])]
+    #[IsGranted('ROLE_ADMIN')]
     public function show(Song $song): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            // Redirige a la página deseada para usuarios que no tienen el rol ROLE_ADMIN
+            return $this->redirectToRoute('app_denied');
+        }
+
         return $this->render('song/show.html.twig', [
             'song' => $song,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_song_edit', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_ADMIN')]
     public function edit(Request $request, Song $song, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            // Redirige a la página deseada para usuarios que no tienen el rol ROLE_ADMIN
+            return $this->redirectToRoute('app_denied');
+        }
+
         $form = $this->createForm(SongType::class, $song);
         $form->handleRequest($request);
 
@@ -109,9 +127,15 @@ class SongController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_song_delete', methods: ['POST'])]
+    #[IsGranted('ROLE_ADMIN')]
     public function delete(Request $request, Song $song, EntityManagerInterface $entityManager): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$song->getId(), $request->request->get('_token'))) {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            // Redirige a la página deseada para usuarios que no tienen el rol ROLE_ADMIN
+            return $this->redirectToRoute('app_denied');
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $song->getId(), $request->request->get('_token'))) {
             $entityManager->remove($song);
             $entityManager->flush();
         }
